+++ rx-location-services/src/main/kotlin/org/neidhardt/rxlocation/services/GoogleLocationService.kt
@@ -11,6 +11,7 @@ import com.google.android.gms.location.LocationServices
 import io.reactivex.BackpressureStrategy
 import io.reactivex.Flowable
 import io.reactivex.Single
+import io.reactivex.android.schedulers.AndroidSchedulers
 import org.neidhardt.rxlocation.exceptions.MissingPermissionCoarseLocation
 import org.neidhardt.rxlocation.exceptions.MissingPermissionFineLocation
 
@@ -43,6 +44,7 @@ class GoogleLocationService(private val context: Context) {
 	 *
 	 * @return single location update
 	 */
+	@Suppress("unused")
 	@SuppressLint("MissingPermission")
 	fun getLastKnowLocationFromDevice(): Single<Location> {