+++ rx-location-services/src/main/kotlin/org/neidhardt/rxlocation/services/PermissionUtils.kt
@@ -3,7 +3,7 @@ package org.neidhardt.rxlocation.services
 import android.Manifest
 import android.content.Context
 import android.content.pm.PackageManager
-import android.support.v4.content.ContextCompat
+import androidx.core.content.ContextCompat
 import org.neidhardt.rxlocation.exceptions.MissingPermissionCoarseLocation
 import org.neidhardt.rxlocation.exceptions.MissingPermissionFineLocation