+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Added
+* AndroidLocationService to get location updates via stock LocationManager
+### Fixed
+* lastKnownBearing not updated when using rotation sensor
+
 ## v0.4.0 - 2020-03-11
 ### Added
 * Bearing calculation using rotation sensor